CountryViewModel: функция getCountries() передает данные в LiveData. CountryFragment слушает state
