+++ app/src/main/java/ru/practicum/android/diploma/country/data/repository/CountriesListGetterImpl.kt
@@ -1,10 +1,10 @@
-package ru.practicum.android.diploma.filter.data.repository
+package ru.practicum.android.diploma.country.data.repository
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import ru.practicum.android.diploma.filter.domain.model.GetCountriesListResult
-import ru.practicum.android.diploma.filter.domain.repository.CountriesListGetter
-import ru.practicum.android.diploma.region.domain.model.Country
+import ru.practicum.android.diploma.country.domain.model.GetCountriesListResult
+import ru.practicum.android.diploma.country.domain.repository.CountriesListGetter
+import ru.practicum.android.diploma.country.domain.model.Country
 import ru.practicum.android.diploma.search.data.source.remote.HHApiRequest
 import ru.practicum.android.diploma.search.data.source.remote.HHApiResponse
 import ru.practicum.android.diploma.search.data.source.remote.NetworkClient
