+++ app/src/main/java/ru/practicum/android/diploma/country/domain/model/Country.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.region.domain.model
+package ru.practicum.android.diploma.country.domain.model
 
 data class Country(
     val id: String,
