+++ app/src/main/java/ru/practicum/android/diploma/country/domain/repository/CountriesListGetter.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.country.domain.repository
+
+import kotlinx.coroutines.flow.Flow
+import ru.practicum.android.diploma.country.domain.model.GetCountriesListResult
+
+interface CountriesListGetter {
+    fun getCountriesList(): Flow<GetCountriesListResult>
+}
