+++ app/src/main/java/ru/practicum/android/diploma/country/domain/usecase/GetCountriesListUseCase.kt
@@ -1,8 +1,8 @@
-package ru.practicum.android.diploma.filter.domain.usecase
+package ru.practicum.android.diploma.country.domain.usecase
 
 import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.filter.domain.model.GetCountriesListResult
-import ru.practicum.android.diploma.filter.domain.repository.CountriesListGetter
+import ru.practicum.android.diploma.country.domain.model.GetCountriesListResult
+import ru.practicum.android.diploma.country.domain.repository.CountriesListGetter
 
 class GetCountriesListUseCase(private val countriesListGetter: CountriesListGetter) {
     fun execute(): Flow<GetCountriesListResult> {
