+++ app/src/main/java/ru/practicum/android/diploma/country/presentation/state/CountryFragmentState.kt
@@ -1,0 +1,9 @@
+package ru.practicum.android.diploma.country.presentation.state
+
+import ru.practicum.android.diploma.country.domain.model.Country
+
+sealed interface CountryFragmentState {
+
+    data class Content(val countries: List<Country>) : CountryFragmentState
+    object Empty : CountryFragmentState
+}
