+++ app/src/main/java/ru/practicum/android/diploma/country/presentation/viewmodel/CountryViewModel.kt
@@ -1,34 +1,38 @@
 package ru.practicum.android.diploma.country.presentation.viewmodel
 
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import ru.practicum.android.diploma.filter.domain.model.GetCountriesListResult
-import ru.practicum.android.diploma.filter.domain.usecase.GetCountriesListUseCase
-import ru.practicum.android.diploma.region.domain.model.Country
+import ru.practicum.android.diploma.country.domain.model.GetCountriesListResult
+import ru.practicum.android.diploma.country.domain.usecase.GetCountriesListUseCase
+import ru.practicum.android.diploma.country.presentation.state.CountryFragmentState
 
 class CountryViewModel(
     private val getCountriesUseCase: GetCountriesListUseCase
 ) : ViewModel() {
 
-    private val countriesList = ArrayList<Country>()
+    private val countryFragmentStateLiveData = MutableLiveData<CountryFragmentState>()
+    fun observeCountryFragmentState(): LiveData<CountryFragmentState> = countryFragmentStateLiveData
 
     fun getCountries() {
         viewModelScope.launch {
             getCountriesUseCase.execute().collect { result ->
                 when (result) {
                     is GetCountriesListResult.Success -> {
-                        countriesList.addAll(result.countries)
-                        renderCountriesAdapter()
+                        renderCountryFragment(CountryFragmentState.Content(result.countries))
                     }
 
-                    is GetCountriesListResult.Problem -> {}
+                    is GetCountriesListResult.Problem -> {
+                        renderCountryFragment(CountryFragmentState.Empty)
+                    }
                 }
             }
         }
     }
 
-    private fun renderCountriesAdapter() {
-
+    private fun renderCountryFragment(state: CountryFragmentState) {
+        countryFragmentStateLiveData.postValue(state)
     }
 }
