+++ app/src/main/java/ru/practicum/android/diploma/di/DataModule.kt
@@ -5,12 +5,12 @@ import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.module
 import ru.practicum.android.diploma.favorites.data.repository.FavoriteVacanciesRepositoryImpl
 import ru.practicum.android.diploma.favorites.domain.repository.FavoriteVacanciesRepository
-import ru.practicum.android.diploma.filter.data.repository.AreasListGetterImpl
-import ru.practicum.android.diploma.filter.data.repository.CountriesListGetterImpl
-import ru.practicum.android.diploma.filter.data.repository.IndustriesListGetterImpl
-import ru.practicum.android.diploma.filter.domain.repository.AreasListGetter
-import ru.practicum.android.diploma.filter.domain.repository.CountriesListGetter
-import ru.practicum.android.diploma.filter.domain.repository.IndustriesListGetter
+import ru.practicum.android.diploma.region.data.repository.AreasListGetterImpl
+import ru.practicum.android.diploma.country.data.repository.CountriesListGetterImpl
+import ru.practicum.android.diploma.industry.data.repository.IndustriesListGetterImpl
+import ru.practicum.android.diploma.region.domain.repository.AreasListGetter
+import ru.practicum.android.diploma.country.domain.repository.CountriesListGetter
+import ru.practicum.android.diploma.industry.domain.repository.IndustriesListGetter
 import ru.practicum.android.diploma.search.data.repository.VacanciesGetterImpl
 import ru.practicum.android.diploma.search.domain.repository.VacanciesGetter
 import ru.practicum.android.diploma.team.data.repository.DevTeamRepositoryImpl
