--- app/src/main/java/ru/practicum/android/diploma/filter/domain/repository/CountriesListGetter.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.filter.domain.repository
-
-import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.filter.domain.model.GetCountriesListResult
-
-interface CountriesListGetter {
-    fun getCountriesList(): Flow<GetCountriesListResult>
-}
