+++ app/src/main/java/ru/practicum/android/diploma/industry/data/repository/IndustriesListGetterImpl.kt
@@ -1,9 +1,9 @@
-package ru.practicum.android.diploma.filter.data.repository
+package ru.practicum.android.diploma.industry.data.repository
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import ru.practicum.android.diploma.filter.domain.model.GetIndustriesListResult
-import ru.practicum.android.diploma.filter.domain.repository.IndustriesListGetter
+import ru.practicum.android.diploma.industry.domain.model.GetIndustriesListResult
+import ru.practicum.android.diploma.industry.domain.repository.IndustriesListGetter
 import ru.practicum.android.diploma.industry.domain.model.Industry
 import ru.practicum.android.diploma.search.data.source.remote.HHApiRequest
 import ru.practicum.android.diploma.search.data.source.remote.HHApiResponse
