+++ app/src/main/java/ru/practicum/android/diploma/industry/domain/model/GetIndustriesListResult.kt
@@ -1,6 +1,4 @@
-package ru.practicum.android.diploma.filter.domain.model
-
-import ru.practicum.android.diploma.industry.domain.model.Industry
+package ru.practicum.android.diploma.industry.domain.model
 
 sealed interface GetIndustriesListResult {
     data class Success(val industries: List<Industry>) : GetIndustriesListResult
