+++ app/src/main/java/ru/practicum/android/diploma/industry/domain/repository/IndustriesListGetter.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.industry.domain.repository
+
+import kotlinx.coroutines.flow.Flow
+import ru.practicum.android.diploma.industry.domain.model.GetIndustriesListResult
+
+interface IndustriesListGetter {
+    fun getIndustriesList(): Flow<GetIndustriesListResult>
+}
