+++ app/src/main/java/ru/practicum/android/diploma/industry/domain/usecase/GetIndustriesListUseCase.kt
@@ -1,8 +1,8 @@
-package ru.practicum.android.diploma.filter.domain.usecase
+package ru.practicum.android.diploma.industry.domain.usecase
 
 import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.filter.domain.model.GetIndustriesListResult
-import ru.practicum.android.diploma.filter.domain.repository.IndustriesListGetter
+import ru.practicum.android.diploma.industry.domain.model.GetIndustriesListResult
+import ru.practicum.android.diploma.industry.domain.repository.IndustriesListGetter
 
 class GetIndustriesListUseCase(private val industriesListGetter: IndustriesListGetter) {
     fun execute(): Flow<GetIndustriesListResult> {
