+++ app/src/main/java/ru/practicum/android/diploma/region/domain/repository/AreasListGetter.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.region.domain.repository
+
+import kotlinx.coroutines.flow.Flow
+import ru.practicum.android.diploma.region.domain.model.GetAreasListResult
+
+interface AreasListGetter {
+    fun getAreasList(countryId: String): Flow<GetAreasListResult>
+}
