+++ app/src/main/java/ru/practicum/android/diploma/region/domain/usecase/GetAreasListUseCase.kt
@@ -1,8 +1,8 @@
-package ru.practicum.android.diploma.filter.domain.usecase
+package ru.practicum.android.diploma.region.domain.usecase
 
 import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.filter.domain.model.GetAreasListResult
-import ru.practicum.android.diploma.filter.domain.repository.AreasListGetter
+import ru.practicum.android.diploma.region.domain.model.GetAreasListResult
+import ru.practicum.android.diploma.region.domain.repository.AreasListGetter
 
 class GetAreasListUseCase(private val areasListGetter: AreasListGetter) {
     fun execute(countryId: String): Flow<GetAreasListResult> {
